move questions & answers controller to admin namespace; in user namespace questions & answers only read; refactoring admin layout
