+++ app/views/admin/layouts/admin.html.erb
@@ -4,8 +4,8 @@
   <title>Admin Panel | TestGuru</title>
   <%= csrf_meta_tags %>
 
-  <%= stylesheer_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-  <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
+  <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
+  <%= javascript_importmap_tags %>
 </head>
 <body>
 <%= render 'admin/nav' %>
